Code cleanup, add celery support (optional)
Code refactored to not only store computed values from memcache, but also to fetch the configured GitHub object from memcache too. This is already an impovement when there are multiple plugins using the same repo (very likely).

The real gains are that the memcache entry is now asynchronously populated so now, the only time GitHub will be accessed from the req/res cycle is if Celery stops working or, the page is accessed before Celery kicks off for the first time.